fix bug textFragment and order css (add tab , color )
